+++ packages/wow-ui/src/components/RadioGroup/index.ts
@@ -1,0 +1,2 @@
+export { default as RadioButton } from "./RadioButton";
+export { default as RadioGroup } from "./RadioGroup";
